fix: default to bitswap-client-only
This restores bitswap setup we had in v1.0.0 where only the client
is initialized by default.
